stackpythonofficialweb-master/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
Django==2.0
django-comments-xtd==2.4.2
django-contrib-comments==1.9.2
django-crispy-forms==1.8.1
django-jet==1.0.8
django-modelcluster==5.0.1
django-recaptcha==2.0.6
django-social-share==1.3.2
django-summernote==0.8.11.6
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
docutils==0.16
draftjs-exporter==2.1.7
feedparser==5.2.1
html5lib==1.0.1
idna==2.8
Pillow==6.2.2
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.8
webencodings==0.5.1
Willow==1.3
